// src/store/modules/content/mutation-types.ts
export enum MutationTypes {
  BENEFITS_LOADING = "BENEFITS_LOADING",
  BENEFITS_LOADED = "BENEFITS_LOADED",
  BRANCHES_LOADING = "BRANCHES_LOADING",
  BRANCHES_LOADED = "BRANCHES_LOADED",
  COMPANY_LOADING = "COMPANY_LOADING",
  COMPANY_LOADED = "COMPANY_LOADED",
  CULTURAL_FITS_LOADING = "CULTURAL_FITS_LOADING",
  CULTURAL_FITS_LOADED = "CULTURAL_FITS_LOADED",
  JOB_REQUIREMENTS_LOADING = "JOB_REQUIREMENTS_LOADING",
  JOB_REQUIREMENTS_LOADED = "JOB_REQUIREMENTS_LOADED",
  JOB_TYPES_LOADING = "JOB_TYPES_LOADING",
  JOB_TYPES_LOADED = "JOB_TYPES_LOADED",
  JOB_POSITIONS_LOADING = "JOB_POSITIONS_LOADING",
  JOB_POSITIONS_LOADED = "JOB_POSITIONS_LOADED",
  LANGUAGES_LOADING = "LANGUAGES_LOADING",
  LANGUAGES_LOADED = "LANGUAGES_LOADED",
  LANGUAGE_LEVELS_LOADING = "LANGUAGE_LEVELS_LOADING",
  LANGUAGE_LEVELS_LOADED = "LANGUAGE_LEVELS_LOADED",
  SKILLS_LOADING = "SKILLS_LOADING",
  SKILLS_LOADED = "SKILLS_LOADED",
  SOFT_SKILLS_LOADING = "SOFT_SKILLS_LOADING",
  SOFT_SKILLS_LOADED = "SOFT_SKILLS_LOADED",
}